// types/skill.ts
import { ListSectionPayload } from './common';

/**
 * ### Sample Rendering
 *
 * ![image](https://user-images.githubusercontent.com/8033320/78029577-cf1b0100-739b-11ea-9c2c-a41acbe9125c.png)
 *
 * @example https://github.com/chhann/resume/blob/master/payload/skill.ts
 */
export interface SkillPayload extends ListSectionPayload {
  /**
   * ### 보유 기술 목록
   */
  skills: SkillItem[];

}

export interface SkillItem {
  /** ### 대분류 */
  category: string;

  /** ### 해당 분류 내 항목들 */
  items: SkillSubItem[];
}

export interface SkillSubItem {
  /** ### 보유 기술 이름 */
  title: string;

  /**
   * ### 보유 기술 수준
   *
   * @value 1: 옅은 회색 뱃지로 나타난다. (light)
   * @value 2: 짙은 회색 뱃지로 나타난다. (sencondary)
   * @value 3: 파란색 뱃지로 나타난다. (primary)
   * @value undefined: 뱃지가 붙지 않는다.
   */
  level?: 1 | 2 | 3;
}